+++ tests/testthat/test-appMetadata-quarto.R
@@ -29,7 +29,7 @@ test_that("inferQuartoInfo correctly detects info when quarto is provided alone"
   expect_equal(quartoInfo$engines, I(c("knitr")))
 })
 
-test_that("inferQuartoInfo extracts info from metadata", {
+test_that("inferQuartoInfo prefers metadata over quarto inspect", {
   metadata <- fakeQuartoMetadata(version = "99.9.9", engines = c("internal-combustion"))
 
   quartoInfo <- inferQuartoInfo(
@@ -43,30 +43,6 @@ test_that("inferQuartoInfo extracts info from metadata", {
   ))
 })
 
-test_that("inferQuartoInfo prefers using metadata over quarto inspect", {
-  skip_if_no_quarto()
-
-  metadata <- fakeQuartoMetadata(version = "99.9.9", engines = c("internal-combustion"))
-
-  quartoInfo <- inferQuartoInfo(
-    appDir = test_path("quarto-website-r"),
-    appPrimaryDoc = NULL,
-    metadata = metadata
-  )
-  expect_equal(quartoInfo$engines, I(c("internal-combustion")))
-})
-
-test_that("inferQuartoInfo returns NULL for non-quarto content", {
-  skip_if_no_quarto()
-
-  quartoInfo <- inferQuartoInfo(
-    appDir = test_path("shinyapp-simple"),
-    appPrimaryDoc = NULL,
-    metadata = list()
-  )
-  expect_null(quartoInfo)
-})
-
 test_that("quartoInspect requires quarto", {
   local_mocked_bindings(quarto_path = function() NULL)
   expect_snapshot(error = TRUE, {
@@ -74,7 +50,7 @@ test_that("quartoInspect requires quarto", {
   })
 })
 
-test_that("quartoInspect identifies on Quarto projects", {
+test_that("quartoInspect identifies Quarto projects", {
   skip_if_no_quarto()
 
   inspect <- quartoInspect(test_path("quarto-website-r"))
@@ -123,9 +99,42 @@ test_that("quartoInspect processes content with filenames containing spaces", {
   expect_equal(inspect$engines, c("markdown"))
 })
 
-test_that("quartoInspect returns NULL on non-quarto Quarto content", {
+test_that("quartoInspect produces an error when a document cannot be inspected", {
   skip_if_no_quarto()
 
-  inspect <- quartoInspect(test_path("shinyapp-simple"))
-  expect_null(inspect)
+  dir <- local_temp_app(list("bad.qmd" = c(
+    "---",
+    "format: unsupported",
+    "---",
+    "this is a document using an unsupported format."
+  )))
+  expect_error(
+    quartoInspect(dir, "bad.qmd"),
+    "Unable to run `quarto inspect` against your content",
+    fixed = TRUE
+  )
+})
+
+test_that("quartoInspect produces an error when a project cannot be inspected", {
+  skip_if_no_quarto()
+
+  dir <- local_temp_app(
+    list(
+      "_quarto.yml" = c(
+        "project:",
+        "  type: unsupported"
+      ),
+      "bad.qmd" = c(
+        "---",
+        "title: bad",
+        "---",
+        "this is a document using an unsupported format."
+      )
+    )
+  )
+  expect_error(
+    quartoInspect(dir, "bad.qmd"),
+    "Unable to run `quarto inspect` against your content",
+    fixed = TRUE
+  )
 })